[step5] refactor method naming
